+++ docs/changelog.rst
@@ -19,6 +19,8 @@ New features
   * New class :class:`aspecd.dataset.DeviceData` for device data.
   * New class :class:`aspecd.analysis.DeviceDataExtraction` for extracting device data from a dataset as a separate dataset. This allows to proceed with the extracted datasets as with any other dataset.
 
+* New parameter ``device_data`` in :class:`aspecd.plotting.Plotter` for plotting device data rather than primary data of a dataset/datasets
+
 
 Version 0.8.4
 =============